+++ packages/aws-amplify-react/__tests__/Auth/ForgotPassword-test.js
@@ -145,4 +145,4 @@ describe('forgotPassword', () => {
             }
         });
     });
-})
+});